backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
black==23.12.1
cachecontrol==0.13.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.101.1
flake8==6.1.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
idna==3.10
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
pyasn1-modules==0.4.2
pyasn1==0.6.1
pycodestyle==2.11.1
pydantic-core==2.33.1
pydantic==2.11.2
pyflakes==3.1.0
pygments==2.19.1
pyinstrument==4.7.3
pyright==1.1.398
python-dotenv==1.1.0
requests==2.32.3
rich==13.9.4
rsa==4.9
scipy==1.13.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy-cockroachdb==2.0.2
sqlalchemy==2.0.40
starlette==0.27.0
tomli==2.2.1
typer[all]==0.9.4
typing-extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.23.2
vulture==2.14
firebase-admin==6.4.0
